+++ source-files/web-components/debug-panel.js
@@ -308,6 +308,17 @@ window.customElements.define('debug-panel', class extends HTMLElement {
         });
     });
 
+    this.shadowRoot.getElementById('eraseCacheButtonId').addEventListener('click', () => {
+      document.getElementById('ircControlsPanel').eraseIrcCache('CACHE')
+        .catch((err) => {
+          console.log(err);
+          let message = err.message || err.toString() || 'Error occurred calling /irc/connect';
+          // show only 1 line
+          message = message.split('\n')[0];
+          document.getElementById('errorPanel').showError(message);
+        });
+    });
+
     this.shadowRoot.getElementById('serverTerminateButtonId').addEventListener('click', () => {
       document.getElementById('ircControlsPanel').webServerTerminate()
         .catch((err) => {